최종 데모를 위해서 SQLite 데이터 확인하는 버튼과 액티비티 추가

+++ app/src/main/java/com/explain/VoiceActivity.java
@@ -128,6 +128,7 @@ public class VoiceActivity extends AppCompatActivity {
         });
     }
 
+
     public void onPause() {
         super.onPause();
         db.deleteData();
@@ -159,7 +160,7 @@ public class VoiceActivity extends AppCompatActivity {
         if(numberOfNoun == 0)
             averageUseCount = 1;
         else
-            averageUseCount = (double) totalCountOfNoun / numberOfNoun;
+            averageUseCount = (double)totalCountOfNoun / (double)numberOfNoun;
 
 
         if(data.get(targetWord) < averageUseCount)
